move resolve into c-gull and remove net feature
