+++ c-gull/Cargo.toml
@@ -17,7 +17,8 @@ libc = { version = "0.2.138", default-features = false }
 c-scape = { path = "../c-scape", version = "0.6.1", default-features = false }
 errno = { version = "0.2.8", default-features = false, optional = true }
 tz-rs = { version = "0.6.11", optional = true }
-printf-compat = { version = "0.1.1", optional = true}
+printf-compat = { version = "0.1.1", optional = true }
+sync-resolve = { version = "0.3.0", optional = true }
 log = { version = "0.4.14", default-features = false, optional = true }
 rustix = { version = "0.36.1", default-features = false, optional = true, features = ["fs", "itoa", "net", "param", "process", "rand", "termios", "thread", "time"] }
 
@@ -27,4 +28,4 @@ libc = "0.2.138"
 [features]
 default = ["threads", "std"]
 threads = ["c-scape/threads"]
-std = ["c-scape/default", "rustix", "log", "printf-compat", "tz-rs", "errno"]
+std = ["c-scape/std", "rustix", "log", "printf-compat", "tz-rs", "errno", "sync-resolve"]
